Move flowchart/state into the family-plugin registry (no in-tree exception)
Answers 'why can't we modernise flowchart/state': we could — the BUILD-3
exception was a scoping choice, not a constraint. The four quirks that
kept it in-tree are now expressed in the plugin contract instead:

- Two kinds, one body: flowchart and state register as two plugins
  sharing one implementation (flowchartFamilyHooks) with the serialized
  header bound per kind; parse/serialize/mutate dispatch by DIAGRAM kind,
  so state diagrams hit the state plugin and keep stateDiagram-v2.
- Error semantics: FamilyPlugin.parse now returns ParseError[] and
  receives canonicalSource, so the legacy parser's throw-on-bad-syntax
  contract survives (no silent opaque fallback for the flagship family).
- SourceMap: new optional buildSourceMap hook; flowchart/state index
  node positions, other families return no map.
- Behavior upgrade: flowchart/state mutation now rebuilds
  canonicalSource like every other family (previously stale), pinned by
  a state-header regression test.

src/agent/flowchart-body.ts now owns the graph serializer, mutator,
parse wrapper, and source-map builder (moved verbatim from
serialize.ts / mutate.ts / parse.ts); the central dispatchers contain
zero per-family branches. The deep modernization — a dedicated
StateBody IR with state-shaped ops — is scoped as TODO BUILD-19,
evidence-gated. Full suite: 1884/1884.
