+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This changelog tracks user-facing changes for **Agentic Mermaid**, a fork of `lu
 
 ### Added
 - **Journey structured mutation** (BUILD-15, the pilot promotion of a source-level family): simple journeys (title/sections/`task: score: actors`) now parse to a typed `JourneyBody`, narrow via `asJourney`, and expose 10 mutation ops (`set_title`, `add_section`, `remove_section`, `set_section_label`, `add_task`, `remove_task`, `set_task_text`, `set_task_score`, `set_task_actors`, `rename_actor`) across the library, CLI (`am mutate`), and MCP Code Mode SDK. Unmodeled journey syntax (accTitle/accDescr, out-of-range scores, header suffixes) still falls back to a lossless opaque body.
-- Family-plugin dispatch consolidation (BUILD-3): parse, serialize, and mutate for sequence, timeline, class, ER, and journey now route through the `FamilyPlugin` registry — adding a structured family is one body module plus one registration. Flowchart/state remain the documented in-tree exception (shared legacy graph body, error semantics, SourceMap). Mutation now rebuilds `canonicalSource` for every plugin family, so mutated sequence diagrams no longer carry stale source.
+- Family-plugin dispatch consolidation (BUILD-3): parse, serialize, and mutate for ALL families — including flowchart/state — now route through the `FamilyPlugin` registry; adding a structured family is one body module plus one registration. Flowchart and state register as two plugins sharing one implementation over the legacy graph body, with the serialized header bound per kind; dispatch is by diagram kind, the plugin contract gained `canonicalSource`/multi-error parse and an optional `buildSourceMap` hook. Mutation now rebuilds `canonicalSource` uniformly, so mutated sequence/flowchart/state diagrams no longer carry stale source.
 - Layout before/after comparison harness (`eval/layout-compare/run.ts`): snapshots the docs corpus + targeted fixtures (SVG, ASCII, perceptual metrics) per git state and emits a side-by-side HTML report with metric deltas and a regression exit code. Geometry tests now also pin subgraph `direction` support — honored even when an inner node links outward, which Mermaid itself does not solve (mermaid-js#2509).
 - **Breaking package identity**: first Agentic Mermaid release is prepared as `agentic-mermaid@0.1.0`; package imports are now `agentic-mermaid` and `agentic-mermaid/agent` while the GitHub repo remains `adewale/beautiful-mermaid`.
 - **Agent-native surface** (`agentic-mermaid/agent` subpath export): a typed editing API for agents and tools.
